Require Bootstrap as a dependency and update assets
Maintaining assets up-to-date may be difficult sometimes when you have a lot of them.
To minimize this work, Bootstrap was added as a runtime dependency
so that its most recent stable version is installed along with sb-admin.
As for the other dependencies, the lack of an official maintained Rails gem makes it impossible to do the same.
Therefore, those were updated to their latest version as of the date of this commit.